Rename config pkg to crud
