+++ teamserver/pkg/crud/Groups.go
@@ -1,4 +1,4 @@
-package config
+package crud
 
 import (
 	"errors"
